updating examples and documentation formatting
