--- docker/aasp/Dockerfile.build
@@ -1,18 +1,0 @@
-ARG GOLANG_VERSION=1.19
-
-FROM golang:${GOLANG_VERSION} AS build
-ENV PROJECT_DIR=/go/src/github.com/microsoft
-WORKDIR ${PROJECT_DIR}/confidential-sidecar-containers
-COPY . ./
-RUN cd tools/get-snp-report && make && mv bin/get-snp-report / && mv bin/get-fake-snp-report / && mv bin/verbose-report /
-
-RUN cd cmd/aasp && CGO_ENABLED=0 GOOS=linux go build -o /aasp -ldflags="-s -w" main.go
-
-
-FROM alpine:3.18.4
-
-RUN apk update && apk --no-cache add curl
-
-COPY --from=build /aasp /get-snp-report /get-fake-snp-report /verbose-report ./bin/
-
-CMD [ "/bin/aasp" ]
