+++ docker/skr/Dockerfile.sample
