+++ examples/skr/aci/aci-arm-template.json
@@ -40,7 +40,7 @@
                   "value": "<optional-loglevel-trace-debug-info-warning-error-fatal-panic>"
                 }
               ],
-              "image": "mcr.microsoft.com/aci/skr:2.5",
+              "image": "mcr.microsoft.com/aci/skr:2.8",
               "resources": {
                 "requests": {
                   "cpu": 1,
@@ -74,7 +74,7 @@
                   "value": "<key-vault-key-identifier>"
                 }
               ],
-              "image": "mcr.microsoft.com/aci/skr:2.5",
+              "image": "mcr.microsoft.com/aci/skr:2.8",
               "resources": {
                 "requests": {
                   "cpu": 0.5,
@@ -120,7 +120,7 @@
                   "value": "eyJrZXlzIjpbeyJlIjoiQVFBQiIsImtleV9vcHMiOlsiZW5jcnlwdCJdLCJraWQiOiJOdmhmdXEyY0NJT0FCOFhSNFhpOVByME5QXzlDZU16V1FHdFdfSEFMel93Iiwia3R5IjoiUlNBIiwibiI6InY5NjVTUm15cDh6Ykc1ZU5GdURDbW1pU2VhSHB1akcyYkNfa2VMU3V6dkRNTE8xV3lyVUp2ZWFhNWJ6TW9PMHBBNDZwWGttYnFIaXNvelZ6cGlORExDbzZkM3o0VHJHTWVGUGYyQVBJTXUtUlNyek41NnF2SFZ5SXI1Y2FXZkhXay1GTVJEd0FlZnlOWVJIa2RZWWtnbUZLNDRoaFVkdGxDQUtFdjVVUXBGWmp2aDRpSTlqVkJkR1lNeUJhS1FMaGpJNVdJaC1RRzZaYTVzU3VPQ0ZNbm11eXV2TjVEZmxwTEZ6NTk1U3MtRW9CSVktTmlsNmxDdHZjR2dSLUlialVZSEFPczVhamFtVHpnZU84a3gzVkNFOUhjeUtteVVac2l5aUY2SURScDJCcHkzTkhUakl6N3Rta3BUSHg3dEhuUnRsZkUyRlV2MEI2aV9RWWxfWkE1USJ9XX0="
                 }
               ],
-              "image": "mcr.microsoft.com/aci/skr:2.5",
+              "image": "mcr.microsoft.com/aci/skr:2.8",
               "resources": {
                 "requests": {
                   "cpu": 1,
