+++ examples/skr/aks/plaintext
