+++ examples/skr/aks/setup-key-mhsm.sh
